requirements.txt
Django>=3.2.4
rasa>=2.7.1
asgiref>=3.3.4
dj-database-url>=0.5.0
django-utils-six>=2.0
gunicorn>=20.1.0
whitenoise>=5.2.0
pandas>=1.2.4
django-filter>=2.4.0
django-widget-tweaks>=1.4.8
django-crispy-forms>=1.12.0
sanic<=21.0.0
xhtml2pdf>=0.2.5
requests>=2.25.1
WeasyPrint>=52.5
python-barcode>=0.13.1